backend/requirements.txt
Flask==3.1.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg[binary]==3.2.13
python-dotenv==1.0.1
PyJWT==2.8.0
Werkzeug==3.1.3
google-auth==2.36.0
redis==5.1.1
rq==1.16.2
openai==1.54.4
PyMuPDF==1.24.14
